abstract loading functionality from App.js
