+++ src/components/main/index.js
@@ -6,3 +6,4 @@ export { default as Menu } from "./Menu";
 export { default as Logo } from "./Logo";
 export { default as TopBar } from "./TopBar";
 export { default as TabData } from "./TabData";
+export { default as LoadContext } from "./LoadContext";
